ai-engine/requirements.txt
fastapi
uvicorn
pydantic
anthropic
qdrant-client
python-dotenv
httpx
beautifulsoup4
